requirements.txt
llama-cpp-python
alive-progress